Add files via upload
This is the result of running robobuilder to create a C++ command based tank drive bot with 4 motors. It also includes the outline for a tank drive ( 4 motor, tank ), raise gear ( button), lower gear(button), climb(button), and autonomous ( auto, first 15 sec ) commands. This code needs to be fleshed out. Motor controller code may need to be swapped out, and nothing exists for the rest of the command other than the outline.