Add files via upload
This is the latest from Robotbuilder as to 2/5/17. It includes just the outline of a C++ Command based, 4 motor, tank drive bot with a gear command, climb command, slow tank command tied to specific buttons, and a placeholder for an autonomous command. This check-in also includes the design file RobotBuilder saves along with your design. In this case it is named SRR 6-rev a.yaml, and is committed in the root of SRR6. 
This is committed after doing a build clean, so once imported into Eclipse, a build will be necessary.