Add files via upload
This is a C++ Command based Eclipse project and source code that allows for Tank drive, Gear, Climber, Slow Tank drive, and autonomous. This is based on SRR6 with modifications made by Eli to fix some button assignment and pneumatic fixes. This is the Skeleton only, and needs to be populated with action code.